Run updates in game loop with fixed time step

+++ PG/animation/AnimationUpdater.cpp
@@ -6,7 +6,7 @@
 namespace PG {
 
 //--------------------------------------------------------
-void AnimationUpdater::updateAnimations(AnimationPtrArray& animations, float dt) const
+void AnimationUpdater::updateAnimations(AnimationPtrArray& animations, double dt) const
 {
 	for (auto& animation : animations)
 	{
