+++ PG/animation/AnimationUpdater.h
@@ -8,7 +8,7 @@ class AnimationPtrArray;
 class AnimationUpdater
 {
 public:
-	void updateAnimations(AnimationPtrArray& animations, float dt) const;
+	void updateAnimations(AnimationPtrArray& animations, double dt) const;
 	
 	void removeCompletedAnimations(AnimationPtrArray& animations) const;
 };
